[refactor] ConversationView SubView들로 분리
