+++ DiverBook_iOS/DiverBook_iOS/Sources/Presentation/ConversationView/View/SubView/QuestionCardCell.swift
@@ -1,0 +1,15 @@
+//
+//  QuestionCardCell.swift
+//  DiverBook_iOS
+//
+//  Created by 배현진 on 4/27/25.
+//
+
+import SwiftUI
+
+struct QuestionCardCell: View {
+    var body: some View {
+        Image("questionCard")
+            .shadow(color: DiveColor.gray4.opacity(0.18), radius: 5, y: 5)
+    }
+}
